+++ CHANGELOG.md
@@ -3,10 +3,14 @@ X.Y.Z (YYYY-MM-DD)
 
 New
 ~~~
+- Added basic default logging Logic and Decorators.
 - Added commenting Feature on Instances.
 - Added Cover Images on Instances view/edit Pages.
 - Added Google Analytics Tag.
 - Added view on Map Feature for Instances.
+
+Chore
+~~~~~
 
 Changes
 ~~~~~~~
@@ -14,6 +18,9 @@ Changes
 - Updated Admins.
 - Updated Event List grouping.
 - Updated Sitemap-generating Logic.
+
+Doc
+~~~
 
 Fix
 ~~~